pkg/api/core: move pkg/api/resource
