+++ pkg/apis/core/helper/resource/BUILD
@@ -9,7 +9,7 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["helpers.go"],
-    importpath = "k8s.io/kubernetes/pkg/api/resource",
+    importpath = "k8s.io/kubernetes/pkg/apis/core/helper/resource",
     deps = [
         "//pkg/apis/core:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
@@ -33,7 +33,7 @@ go_test(
     name = "go_default_test",
     srcs = ["helpers_test.go"],
     embed = [":go_default_library"],
-    importpath = "k8s.io/kubernetes/pkg/api/resource",
+    importpath = "k8s.io/kubernetes/pkg/apis/core/helper/resource",
     deps = [
         "//pkg/apis/core:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
