+++ pkg/kubectl/.import-restrictions
@@ -7,7 +7,7 @@
                 "k8s.io/kubernetes/pkg/api/legacyscheme",
                 "k8s.io/kubernetes/pkg/apis/core/helper/pod",
                 "k8s.io/kubernetes/pkg/api/ref",
-                "k8s.io/kubernetes/pkg/api/resource",
+                "k8s.io/kubernetes/pkg/apis/core/helper/resource",
                 "k8s.io/kubernetes/pkg/apis/core/helper/service",
                 "k8s.io/kubernetes/pkg/apis/core/v1/helper/pod",
                 "k8s.io/kubernetes/pkg/apis/core/v1/helper/service",
